requirements.txt
Django==4.2.16
djangorestframework==3.14.0
django-cors-headers==4.3.1
psycopg2-binary==2.9.9
drf-yasg==1.21.7
python-decouple==3.8
django-filter==23.5
Pillow==10.1.0
drf-api-tracking==1.8.0
python-decouple==3.8
djangorestframework-simplejwt==5.3.0
